==> Spotify Hit Prediction/Preprocessing/Data Wrangling/4_Spotipy - Add Features.ipynb <==
In this Notebook I will get the Audio Analysis features for the billboard songs

# get api key from your .env file

import os

from dotenv import load_dotenv

from pathlib import Path

 

dotenv_path = Path('./SpotifyAPIKey.env')

load_dotenv(dotenv_path=dotenv_path)


#load_dotenv()

API_CLIENT_ID = os.getenv('SPOTIFY_CLIENT_ID')
API_CLIENT_SECRET = os.getenv('SPOTIFY_CLIENT_SECRET')
 

import pandas as pd
import spotipy
from spotipy.oauth2 import SpotifyClientCredentials
import numpy as np


#https://spotipy.readthedocs.io/en/2.19.0/

client_credentials_manager = SpotifyClientCredentials(client_id=API_CLIENT_ID,client_secret=API_CLIENT_SECRET)
sp = spotipy.Spotify(client_credentials_manager=client_credentials_manager)

#sp.audio_features({'URI'}) gives us a dictionary within a list of the features and their values
#testing how this works

def GetAudioFeatures(URI):
    features = sp.audio_features(URI)
    #df_features = pd.DataFrame.from_dict(features)
    #dataframe = dataframe.merge(df_features, how='inner', left_on='URI', right_on='id')
    return features[0]
 


dict1_ = sp.audio_features('5c1QJAiC3773EwiZLwHf5x')

GetAudioFeatures('5c1QJAiC3773EwiZLwHf5x')

df_t1 = pd.DataFrame.from_dict(dict1_)

df_t1.head()

d = {'URI': ['5c1QJAiC3773EwiZLwHf5x', '5eCkuGPZNl4mISFNPgr3Dd'], 'Artist': ['name1', 'name2']}

df2_test = pd.DataFrame(data=d)

df2_test.head()

df2_test = GetAudioFeatures(df2_test['URI'], df2_test)

df2_test.head()

Now that I have checked the code works, I will import my dataframes and run the funtion on them. 

billboard_songs = pd.read_csv('BillboardWithURI.csv', usecols=[1,2,3])

flop_songs = pd.read_csv('flop_songs.csv', usecols=[1,2,3])

flp = pd.read_csv('flop_songs.csv', usecols=[1,2,3])

flp.info()

billboard_songs.head()

billboard_songs.info()

df = billboard_songs['URI'].map(lambda x : GetAudioFeatures(x))

df.head()

df = df.apply(pd.Series)

df.head()

billboard_songs = billboard_songs.merge(df, how='inner', left_on='URI', right_on='id')

billboard_songs.info()

duplicates = billboard_songs[billboard_songs.duplicated(subset='URI')]
duplicates

flop_songs.head()

df1 = flop_songs['URI'].map(lambda x : GetAudioFeatures(x))

df1 = df1.apply(pd.Series)

df1.head()

flop_songs = flop_songs.merge(df1, how='inner', left_on='URI', right_on='id')

flop_songs.info()

dups = billboard_songs[billboard_songs.duplicated(subset='URI')]
dups

8 flop songs did not return any features..

billboard_songs['Target'] = 1

billboard_songs.head()

flop_songs['Target'] = 0

flop_songs.head()

dupies = flop_songs[flop_songs.duplicated(subset='URI')]
dupies

df_con = pd.concat([billboard_songs, flop_songs])

df_con.head()

df_con.info()

df_con = df_con.drop(columns = ['type', 'id','uri','track_href', 'analysis_url'])

df_con.info()

df_con.to_csv('FullDataset.csv')
